requirements.txt
feedparser==6.0.2
tqdm
scikit-learn
beautifulsoup4
plotly
pandas
spacy
umap-learn